Changed how the emoji is verified.
When using custom emoji the line where the code checks if the emoji reacted is same as the emoji in config the script stops resulting in no new ticket creation.
To rectify this I have added a new variable `emoji` which is the emoji name appended with emoji ID so that the script is not stopped. In config, the panel reacts with emoji only if we put the name followed by its ID (eg: :name:123456576875) and in the IF statement the verification stops because of `r.emoji.name` gives only and `r.emoji.id` gives only the ID which is obviously not same as the `config.panel.reaction`.